chore (refs T22614): move DpPager to client/js.
