+++ demosplan/DemosPlanStatementBundle/Resources/client/js/components/assessmentTable/DpTable.vue
@@ -239,7 +239,7 @@ import { CleanHtml } from 'demosplan-ui/directives'
 import DpAssessmentTableCard from '@DemosPlanStatementBundle/components/assessmentTable/DpAssessmentTableCard'
 import DpExportModal from '@DemosPlanStatementBundle/components/assessmentTable/DpExportModal'
 import { DpLoading } from 'demosplan-ui/components'
-import DpPager from '@DemosPlanCoreBundle/components/DpPager'
+import DpPager from '@DpJs/components/core/DpPager'
 import { handleResponseMessages } from '@DemosPlanCoreBundle/plugins/DpApi'
 import { scrollTo } from 'vue-scrollto'
 import Stickier from '@DpJs/lib/Stickier'
